Rework extraction of equipment and accessory
