+++ config/splat.us.dra.yaml
@@ -38,8 +38,8 @@ segments:
       - [0x2EE8, data]
       - [0x3C40, data]
       - [0x4A00, data]
-      - [0x4B04, equipment, equipments]
-      - [0x7718, accessory, accessory_data]
+      - [0x4B04, assets, equipment]
+      - [0x7718, assets, accessory]
       - [0x8258, data]
       - [0x8900, data]
       - [0xCEB0, data]
