+++ tools/splat_ext/equipment_config.json
@@ -1,0 +1,31 @@
+    {
+        "name_addr": "str_ptr",
+        "desc_addr": "str_ptr",
+        "attack": "s16",
+        "defense": "s16",
+        "element": "u16",
+        "itemCategory": "u8",
+        "weaponId": "u8",
+        "palette": "u8",
+        "unk11": "u8",
+        "playerAnim": "u8",
+        "unk13": "u8",
+        "unk14": "u8",
+        "lockDuration": "u8",
+        "chainLimit": "u8",
+        "unk17": "u8",
+        "specialMove": "u8",
+        "isConsumable": "bool",
+        "enemyInvincibilityFrames": "u8",
+        "unk1B": "u8",
+        "unk1C": "u32",
+        "unk20": "u32",
+        "mpUsage": "s16",
+        "stunFrames": "s16",
+        "hitType": "s16",
+        "hitEffect": "s16",
+        "icon": "s16",
+        "iconPalette": "s16",
+        "criticalRate": "s16",
+        "unk32": "s16"
+    }
